+++ colanderalchemy/schema.py
@@ -82,7 +82,9 @@ class SQLAlchemySchemaNode(colander.SchemaNode):
 
            Incompatible with :attr:`includes`. Default: None.
         overrides
-           XXX Add something.
+            A dict-like structure that consists of schema attributes to
+            override imperatively. Values provides as part of :attr:`overrides`
+            will take precendence over all others.
         unknown
            Represents the `unknown` argument passed to
            :class:`colander.Mapping`.
@@ -177,7 +179,9 @@ class SQLAlchemySchemaNode(colander.SchemaNode):
             A given :class:`sqlalchemy.orm.properties.ColumnProperty`
             instance that represents the column being mapped.
         overrides
-            XXX Add something.
+            A dict-like structure that consists of schema attributes to
+            override imperatively. Values provides as part of :attr:`overrides`
+            will take precendence over all others.
         """
 
         # The name of the SchemaNode is the ColumnProperty key.
@@ -334,6 +338,13 @@ class SQLAlchemySchemaNode(colander.SchemaNode):
     def get_schema_from_relationship(self, prop, overrides):
         """ Build and return a :class:`colander.SchemaNode` for a relationship.
 
+        The mapping process will translate one-to-many and many-to-many
+        relationships from SQLAlchemy into a ``Sequence`` of ``Mapping`` nodes
+        in Colander, and translate one-to-one and many-to-one relationships
+        into a ``Mapping`` node in Colander.  The related class involved in the
+        relationship will be recursively mapped by ColanderAlchemy as part of
+        this process, following the same mapping process.
+
         This method uses information stored in the relationship within
         the ``info`` that was passed to the relationship on creation.
         This means that ``Colander`` options can be specified
@@ -341,13 +352,22 @@ class SQLAlchemySchemaNode(colander.SchemaNode):
         argument that you can pass to
         :meth:`sqlalchemy.orm.relationship`.
 
+        For all relationships, the settings will only be applied to the outer
+        Sequence or Mapping. To customise the inner schema node, create the
+        attribute ``__colanderalchemy_config__`` on the related model with a
+        dict-like structure corresponding to the Colander options that should
+        be customised.
+
         Arguments/Keywords
 
         prop
             A given :class:`sqlalchemy.orm.properties.RelationshipProperty`
             instance that represents the relationship being mapped.
         overrides
-            XXX Add something.
+            A dict-like structure that consists of schema attributes to
+            override imperatively. Values provides as part of :attr:`overrides`
+            will take precendence over all others.  Example keys include
+            ``children``, ``includes``, ``excludes``, ``overrides``.
         """
 
         # The name of the SchemaNode is the ColumnProperty key.
@@ -439,11 +459,11 @@ class SQLAlchemySchemaNode(colander.SchemaNode):
         # Add default values for missing parameters.
         if prop.innerjoin:
             #Inner joined relationships imply it is mandatory
-            kwargs["missing"] = required
+            missing = required
         else:
             #Any other join is thus optional
-            kwargs["missing"] = drop
-
+            missing = drop
+        kwargs['missing'] = missing
 
         kwargs.update(declarative_overrides)
         kwargs.update(overrides)
@@ -457,10 +477,11 @@ class SQLAlchemySchemaNode(colander.SchemaNode):
                 return SchemaNode(Mapping(), *children, **kwargs)
 
         node = SQLAlchemySchemaNode(class_,
+                                    name=name,
                                     includes=includes,
                                     excludes=excludes,
                                     overrides=rel_overrides,
-                                    **kwargs)
+                                    missing=missing)
 
         if prop.uselist:
             node = SchemaNode(Sequence(), node, **kwargs)